+++ src/main/java/eevee/gui/MainWindow.java
@@ -1,7 +1,9 @@
-package eevee;
+package eevee.gui;
 
+import eevee.Eevee;
+import eevee.EeveeException;
+import eevee.gui.DialogBox;
 import javafx.fxml.FXML;
-import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
