+++ src/main/java/eevee/task/Event.java
@@ -1,4 +1,7 @@
-package eevee;
+package eevee.task;
+
+import eevee.EeveeException;
+import eevee.task.Task;
 
 import java.time.DateTimeException;
 import java.time.LocalDate;
