+++ src/main/java/eevee/task/TaskList.java
@@ -1,4 +1,6 @@
-package eevee;
+package eevee.task;
+
+import eevee.EeveeException;
 
 import java.util.ArrayList;
